+++ src/components/ogre/widgets/MeshInfoProvider.cpp
@@ -116,7 +116,7 @@ size_t MeshInfoProvider::calcUniqueVertexCount(const Ogre::VertexData& data)
 int MeshInfoProvider::getVertexSize(const Ogre::VertexData* data)
 {
 	int size = 0;
-	for (size_t i = 0; i < data->vertexDeclaration->getMaxSource() + 1; i++) {
+	for (unsigned short i = 0; i < data->vertexDeclaration->getMaxSource() + 1; i++) {
 		size += data->vertexDeclaration->getVertexSize(i);
 	}
 	return size;
